Update & Load Path
I have managed to get the update path function to work. Everytime user clicks to add a point, the update function will run which will send latlang data to database using SQL commands.

I am just about to get the load path function to work, it is only a matter of putting it into an array and loading that array